libs/containerv: implement a pid1 daemon (#92)
* libs/containerv/windows: wire in the pid1 creation

* libs/containerv: pid1d added

* libs/containerv: more pid1 support

* libs/containerv/windows: better vmdisk infrastructure

* libs/containerv/windows: updated to emit guest-appropriate defaults

* libs/containerv/windows: Linux guests use HCSProcessParameters.CommandArgs (array) instead of relying on a Windows-style quoted CommandLine string

* libs/containerv/windows: routed VM guest network setup through pid1d instead of trying to run HCS processes directly.

* libs/gracht: update to newest

* many: more windows improvements, gateway/dns support for windows containers

* many: improvements, support for network configuration

* many: cleanup, proper windows recipes

* many: last batch of changes to improve windows support and restructure code

* cmake: fix the zstd library name of windows

* many: build fixes and compiler improvements

* libs/containerv: more work on LCOW

* many: build fixes

* libs/containerv: more OCI stuff

* many: restructure containerv-disk

* many: improve support for LCOW and WCOW

* libs/containerv: remove the windows VM backend

* server.c: fix storage warning

* many: support lcow uvm

* libs/containerv: fix build issues on windows

* many: more build fixes

* libs/containerv/windows: remove file we dont use anymore

* libs/containerv/windows: cleanup unused things

* libs/containerv/windows: remove leftovers

* libs/containerv/windows: vmcompute is already dynamically loaded

* many: add the windows resolver

* meson: remove usage of open_memstream

+++ daemons/served/utils/proofs.c
@@ -83,7 +83,7 @@ static int __verify_signature_against_cert(struct store_proof_publisher* proof)
     EVP_PKEY* pubkey;
     int       status;
 
-    keybio = BIO_new_mem_buf(g_certAuthority, strlen(g_certAuthority));
+    keybio = BIO_new_mem_buf(g_certAuthority, (int)strlen(g_certAuthority));
     if (keybio == NULL) {
         VLOG_ERROR("served", "__verify_signature_against_cert: failed to create memory reader\n");
         return -1;
