+++ libs/chefclient/storage/bashupload.c
@@ -57,13 +57,13 @@ static const char* g_upload0 = "uploading [";
 
 static void __update_progress(struct upload_context* downloadContext)
 {
-    char progressBuffer[32];
-    char totalSizeBuffer[32];
-    char buffer[64] = { 0 };
-    int  index = 0;
-    int  percent;
+    char   progressBuffer[32];
+    char   totalSizeBuffer[32];
+    char   buffer[64] = { 0 };
+    size_t index = 0;
+    int    percent;
 
-    percent = (downloadContext->bytes_uploaded * 100) / downloadContext->bytes_total;
+    percent = (int)((downloadContext->bytes_uploaded * 100) / downloadContext->bytes_total);
 
     // print a fancy progress bar with percentage, upload progress and a moving
     // bar being filled
